Fix the call logic of the gd32f4xx_appinit.C and gd32f4xx_bringup.C of the GD32 series microcontroller in the project, which is unified with the STM32 series chip in the warehouse.
